refactor(inference): share the bounded config reader, centralize the resize pixel budget, and table the rescale
Moved the bounded configuration-file reader into a shared module. Applied one pixel-budget policy through smart_resize and the resizer, corrected the coefficient-table documentation, and replaced per-pixel rescaling with a 256-entry table with identical results.
